# Vary table size per tree.