--- a/src/lsm/forest.zig
+++ b/src/lsm/forest.zig
@@ -67,6 +67,7 @@ pub fn ForestType(comptime Storage: type, comptime groove_config: anytype) type 
             open,
         };
 
+        pub const groove_config = groove_config;
         pub const GroovesOptions = _GroovesOptions;
 
         join_op: ?JoinOp = null,