--- a/src/lsm/segmented_array.zig
+++ b/src/lsm/segmented_array.zig
@@ -1241,6 +1241,7 @@ pub fn run_tests(seed: u64, comptime options: Options) !void {
         Key.sentinel_key,
         Key.tombstone,
         Key.tombstone_from_key,
+        1, // Doesn't matter for this test.
         .general,
     ));