--- a/src/state_machine/auditor.zig
+++ b/src/state_machine/auditor.zig
@@ -17,6 +17,7 @@ const PriorityQueue = @import("../testing/priority_queue.zig").PriorityQueue;
 const Storage = @import("../testing/storage.zig").Storage;
 const StateMachine = @import("../state_machine.zig").StateMachineType(Storage, .{
     .message_body_size_max = constants.message_body_size_max,
+    .lsm_batch_multiple = constants.lsm_batch_multiple,
 });
 
 pub const CreateAccountResultSet = std.enums.EnumSet(tb.CreateAccountResult);