Implements https://github.com/tigerbeetledb/tigerbeetle/issues/271.

The diff isn't as bad as it looks. The first commit passes `value_count_max` into table.zig rather than table_mutable.zig so that it can be used for layout, which requires a lot of trivial changes and renaming some conflicting variables. The second commit actually changes the layout - the math becomes much simpler. 

Has no effect on performance in the current benchmark because all the indexes stressed are append-only. Since we don't coalesce non-overlapping tables, append-only index tables were already accidentally batch-sized.

fuzz_lsm_forest and fuzz_lsm_tree pass (I fixed https://github.com/tigerbeetledb/tigerbeetle/issues/486 while I was in there).

## Pre-merge checklist

Performance:

* [x] Compare `zig benchmark` on linux before and after this pr.
``` sh
# benchmark results before
1223 batches in 79.35 s
load offered = 1000000 tx/s
load accepted = 126031 tx/s
batch latency p00 = 5 ms
batch latency p10 = 16 ms
batch latency p20 = 16 ms
batch latency p30 = 16 ms
batch latency p40 = 16 ms
batch latency p50 = 17 ms
batch latency p60 = 81 ms
batch latency p70 = 82 ms
batch latency p80 = 113 ms
batch latency p90 = 147 ms
batch latency p100 = 2286 ms
transfer latency p00 = 5 ms
transfer latency p10 = 5175 ms
transfer latency p20 = 11008 ms
transfer latency p30 = 17640 ms
transfer latency p40 = 24744 ms
transfer latency p50 = 31812 ms
transfer latency p60 = 38970 ms
transfer latency p70 = 46053 ms
transfer latency p80 = 53192 ms
transfer latency p90 = 62166 ms
transfer latency p100 = 69303 ms
    
# benchmark results after
1223 batches in 77.24 s
load offered = 1000000 tx/s
load accepted = 129471 tx/s
batch latency p00 = 5 ms
batch latency p10 = 16 ms
batch latency p20 = 16 ms
batch latency p30 = 16 ms
batch latency p40 = 16 ms
batch latency p50 = 17 ms
batch latency p60 = 77 ms
batch latency p70 = 82 ms
batch latency p80 = 112 ms
batch latency p90 = 147 ms
batch latency p100 = 467 ms
transfer latency p00 = 5 ms
transfer latency p10 = 5194 ms
transfer latency p20 = 10961 ms
transfer latency p30 = 17508 ms
transfer latency p40 = 24600 ms
transfer latency p50 = 31669 ms
transfer latency p60 = 38821 ms
transfer latency p70 = 45858 ms
transfer latency p80 = 52934 ms
transfer latency p90 = 60115 ms
transfer latency p100 = 67203 ms
```
OR
* [ ] I am very sure this PR could not affect performance.