@jamii 2023-02-22
With the bits of transfer.id flipped to ensure that at least one index actually compacts:

```
# main
1223 batches in 722.16 s
load offered = 1000000 tx/s
load accepted = 13847 tx/s
batch latency p00 = 5 ms
batch latency p10 = 21 ms
batch latency p20 = 21 ms
batch latency p30 = 22 ms
batch latency p40 = 26 ms
batch latency p50 = 32 ms
batch latency p60 = 86 ms
batch latency p70 = 92 ms
batch latency p80 = 1029 ms
batch latency p90 = 2317 ms
batch latency p100 = 7089 ms
transfer latency p00 = 5 ms
transfer latency p10 = 10520 ms
transfer latency p20 = 37360 ms
transfer latency p30 = 79495 ms
transfer latency p40 = 138884 ms
transfer latency p50 = 211660 ms
transfer latency p60 = 301902 ms
transfer latency p70 = 407787 ms
transfer latency p80 = 542298 ms
transfer latency p90 = 640523 ms
transfer latency p100 = 712116 ms

# this branch
1223 batches in 121.09 s
load offered = 1000000 tx/s
load accepted = 82583 tx/s
batch latency p00 = 5 ms
batch latency p10 = 26 ms
batch latency p20 = 27 ms
batch latency p30 = 32 ms
batch latency p40 = 37 ms
batch latency p50 = 46 ms
batch latency p60 = 112 ms
batch latency p70 = 142 ms
batch latency p80 = 173 ms
batch latency p90 = 215 ms
batch latency p100 = 3592 ms
transfer latency p00 = 5 ms
transfer latency p10 = 7526 ms
transfer latency p20 = 16638 ms
transfer latency p30 = 27517 ms
transfer latency p40 = 39246 ms
transfer latency p50 = 51056 ms
transfer latency p60 = 63112 ms
transfer latency p70 = 75241 ms
transfer latency p80 = 87654 ms
transfer latency p90 = 101655 ms
transfer latency p100 = 111055 ms
```

So... compaction is pretty slow and we haven't really been measuring it up until now.