@sentientwaffle 2023-02-22
Afaik we are still planning to use larger tables at the last level of each tree so that the total tree capacity is not reduced significantly. Did you want to put that in a separate PR?