@jamii 2023-02-22
> Afaik we are still planning to use larger tables at the last level of each tree so that the total tree capacity is not reduced significantly. Did you want to put that in a separate PR?

Yeah, there's some open design problems there. Eg do we just do level 7 of the tree (in which case our metadata costs will explode before we get there)? Or do we do the lowest occupied level (in which case what do we do when the level below starts to fill - do we have to move all the tables down a level?).

I'm also concerned that compacting small tables into big tables is expensive, as we just saw demonstrated above. For an append-only index with 4kb tables (8191 tx/batch * 4 batches * 128 bytes), every 4kb of data inserted will eventually trigger a compaction into the bottom level, giving us 64mb/4kb = 16384x write amplification.