@jamii 2023-02-22
Capacity on disk is still pretty reasonable atm - we can hold (8^7 + 8^6 + 8^5 + 8^4 + 8^3 + 8^2 + 8^1) * 8191 * 4 = 78526920416 transfers. But the metadata cost for that will be high. Not just in memory, but in the time it takes to scan TableInfos.

`value_count_max` is pretty low - we're spending on the order of kbs per tree for table_mutable/immutable. We could reduce the metadata cost by increasing `lsm_batch_multiple`.